data: Move consumed replacement masks
Hunk navigation filters batch-owned fragments before caching the selected
line-level view for later commands.

The same path also owns consumed replacement-mask details, even though
those masks are part of consumed-selection metadata rather than hunk
navigation.

This commit begins separating that metadata concern by moving
replacement-mask filtering into data.consumed_replacement_masks and
importing it from hunk tracking.

Subsequent commits will add direct tests for the mask filter and guard hunk
tracking from consumed-selection metadata reads.
