+++ src/git_stage_batch/data/hunk_tracking.py
@@ -20,7 +20,6 @@ from ..core.models import (
     BinaryFileChange,
     GitlinkChange,
     LineLevelChange,
-    LineEntry,
     RenameChange,
     TextFileDeletionChange,
 )
@@ -32,9 +31,9 @@ from ..core.line_selection import write_line_ids_file
 from ..core.line_identity import preserve_line_ids_from_previous_view
 from ..exceptions import CommandError, NoMoreHunks, exit_with_error
 from ..i18n import _
-from .consumed_selections import read_consumed_file_metadata
 from .auto_advance import resolve_auto_advance
 from . import change_freshness as _change_freshness
+from . import consumed_replacement_masks as _consumed_replacement_masks
 from . import file_hunk_display as _file_hunk_display
 from . import live_diff as _live_diff
 from .selected_change import store as _selected_store
@@ -190,7 +189,9 @@ def apply_line_level_batch_filter_to_cached_hunk() -> bool:
     if should_skip:
         return True
 
-    filtered_line_changes = _filter_consumed_replacement_masks(filtered_line_changes)
+    filtered_line_changes = _consumed_replacement_masks.filter_consumed_replacement_masks(
+        filtered_line_changes
+    )
     if filtered_line_changes is None:
         return True
 
@@ -207,64 +208,6 @@ def apply_line_level_batch_filter_to_cached_hunk() -> bool:
     )
 
     return False
-
-
-def _filter_consumed_replacement_masks(
-    line_changes: LineLevelChange,
-) -> LineLevelChange | None:
-    """Hide synthetic replacement runs created by `include --line --as`."""
-    file_metadata = read_consumed_file_metadata(line_changes.path)
-    replacement_masks = file_metadata.get("replacement_masks", []) if file_metadata else []
-    if not replacement_masks:
-        return line_changes
-
-    normalized_masks: set[tuple[tuple[str, str], ...]] = set()
-    for mask in replacement_masks:
-        deleted_signature = tuple(("-", text) for text in mask.get("deleted_lines", []))
-        added_signature = tuple(("+", text) for text in mask.get("added_lines", []))
-        full_signature = deleted_signature + added_signature
-        if full_signature:
-            normalized_masks.add(full_signature)
-        if deleted_signature:
-            normalized_masks.add(deleted_signature)
-        if added_signature:
-            normalized_masks.add(added_signature)
-
-    filtered_lines = []
-    changed_run: list[LineEntry] = []
-
-    def flush_changed_run() -> None:
-        nonlocal changed_run
-        if not changed_run:
-            return
-        run_signature = tuple(
-            (line.kind, line.display_text())
-            for line in changed_run
-            if line.kind in ("+", "-")
-        )
-        if run_signature not in normalized_masks:
-            filtered_lines.extend(changed_run)
-        changed_run = []
-
-    for line_entry in line_changes.lines:
-        if line_entry.kind in ("+", "-"):
-            changed_run.append(line_entry)
-            continue
-        flush_changed_run()
-        filtered_lines.append(line_entry)
-
-    flush_changed_run()
-
-    has_changes_after_filter = any(line.kind in ("+", "-") for line in filtered_lines)
-    if not has_changes_after_filter:
-        return None
-
-    return LineLevelChange(
-        path=line_changes.path,
-        header=line_changes.header,
-        lines=filtered_lines,
-    )
-
 
 def fetch_next_change() -> Union[LineLevelChange, BinaryFileChange, GitlinkChange, RenameChange, TextFileDeletionChange]:
     """Find the next hunk or binary file that isn't blocked and cache it as selected.
